A single merge works now with rearranging files and folders, so a new insert/delete can be done. Removed metaData file from all the indexes and set the front block code size to be a constant 8. Upgraded to BIS the redundant RandomAccessFile usages.

+++ src/main/java/dynamic_index/IndexRemover.java
@@ -6,16 +6,27 @@ import java.nio.file.DirectoryNotEmptyException;
 import java.nio.file.Files;
 import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
+import java.util.HashSet;
+import java.util.Set;
 
-/**
- * Deletes an index directory (actually any directory, recursively)
- */
+
 public class IndexRemover {
 
-    /**
-     * Delete all index files by removing the given directory
-     */
-    public void removeIndex(String dir) {
+//    private final Set<String> afterMergeExceptionFiles = new HashSet<>();
+    private boolean useExceptions = false;
+
+    public IndexRemover(){
+
+    }
+
+    void removeFilesAfterMerge(String dir){
+        useExceptions = true;
+        File file = new File(dir);
+        deleteFileOrDirectory(file);
+    }
+
+    public void removeAllIndexFiles(String dir) {
+        useExceptions = false;
         File file = new File(dir);
         deleteFileOrDirectory(file);
     }
@@ -34,15 +45,21 @@ public class IndexRemover {
 
     private void deleteDirectory(File toDelete) {
         try {
-            File[] childFiles = toDelete.listFiles();
-            if (childFiles != null) {
-                if (childFiles.length == 0) { //Directory is empty. Proceed for deletion
-                    singleDelete(toDelete);
-                } else {
-                    for (File childFilePath : childFiles) {
-                        deleteFileOrDirectory(childFilePath);
+            // not deleting the merged index directory and files
+            if(useExceptions && !toDelete.getName().equals(Statics.MERGED_INDEX_DIRECTORY)){
+                File[] childFiles = toDelete.listFiles();
+                if (childFiles != null) {
+                    if (childFiles.length == 0) { //Directory is empty. Proceed for deletion
+                        singleDelete(toDelete);
+                    } else {
+                        for (File childFilePath : childFiles) {
+                            deleteFileOrDirectory(childFilePath);
+                        }
+                        // not deleting the indexes directory itself after deleting its contents
+                        if(useExceptions && !toDelete.getName().equals(Statics.INDEXES_DIR_NAME)){
+                            deleteDirectory(toDelete); // calling again, now should be empty
+                        }
                     }
-                    deleteDirectory(toDelete); // calling again, now should be empty
                 }
             }
         } catch (NullPointerException e) {
@@ -52,17 +69,21 @@ public class IndexRemover {
 
     private void singleDelete(File singleFileToDelete) {
         Path dirPathToDelete = singleFileToDelete.toPath();
-        try {
-            Files.delete(dirPathToDelete);
-        } catch (NoSuchFileException x) {
-            System.err.format("%s: no such file or directory%n", dirPathToDelete.toString());
-        } catch (DirectoryNotEmptyException x) {
-            System.err.format("%s not empty%n", dirPathToDelete.toString());
-        } catch (IOException x) {
-            x.printStackTrace();
-            System.out.println("FILE NAME: " + singleFileToDelete.getName());
-            System.err.println(x.getMessage());
-            System.exit(3);
+        if(!singleFileToDelete.getName().equals(Statics.INVALIDATION_VECTOR_FILENAME)){
+            try {
+                Files.delete(dirPathToDelete);
+            } catch (NoSuchFileException x) {
+                System.err.format("%s: no such file or directory%n", dirPathToDelete.toString());
+            } catch (DirectoryNotEmptyException x) {
+                System.err.format("%s not empty%n", dirPathToDelete.toString());
+            } catch (IOException x) {
+                x.printStackTrace();
+                System.out.println("FILE NAME: " + singleFileToDelete.getName());
+                System.err.println(x.getMessage());
+                System.exit(3);
+            }
         }
     }
+
+
 }
